+++ ev3dev/_platform/ev3.py
@@ -26,8 +26,7 @@
 An assortment of classes modeling specific features of the EV3 brick.
 """
 
-from .core import *
-
+from ev3dev._button import ButtonEVIO
 
 OUTPUT_A = 'outA'
 OUTPUT_B = 'outB'
