+++ ev3dev/_platform/evb.py
@@ -3,8 +3,7 @@
 An assortment of classes modeling specific features of the EVB.
 """
 
-from .core import *
-
+from ev3dev._button import ButtonEVIO
 
 OUTPUT_A = 'outA'
 OUTPUT_B = 'outB'
